requirements.txt
langchain
langchain-community
python-dotenv
openai

-e .